Small Update
I've added some debugging. Now I can log the message specific to the context: info, warning or errors, all signatured with the proper tag: [LOG], [WARN], [ERR]. What's more, I've optimized the code for generating the MatchTable and replaced the selecting of the object's with the built-in GetMousePosition(), before modified with the offset. Also, cloned and used raygui.