Textures and general tweaks
I've moved the table area from top-left (0,0) to the center of the renderer. What's more, I changed the difficulty; now the grid sizes for each difficulty is: EASY(16 cards, 4^2), NORMAL(36 cards, 6^2), HARD(64 cards, 8^2). I've also changed the resolution to fir all the cards on the screen without sacrificing their size on the screen with scaling. Makefile was changed to not compile and link directly from the source files, but firstly, to .obj files and then to the final executable. Lastly, I've started to work with some graphics. I need to settle down on if I want to use the texture atlas, build another menager (please no) or code all the graphics by myself. That will be decided in the future.