Huge Memo commit
Added new resources: downloaded the Kenney's audio pack, current font (Peaberry), started using rGuiIcons and rGuiLayout for the GUI layout. What's more, I've finally added the settings, and the option to manipulate some of the behaviours of the game. Virtual resolution was changed from 400x300 to 400x296, to match the rGuiLayout grid placement.