Slight changes
This commit is purely just to prepare the project to add another library. From what's new: added the PlayerData.cpp file and moved the definitions of the methods to this source file. What's more, I've modified the behaviour of settings scene: arrows are now responding to the current section, and if there isn't any less / more sections, the widgets goes gray. Also, added another section: settings for IO.